'/lib' renamed to '/cli'
